+++ graphite_beacon/core.py
@@ -1,18 +1,16 @@
+import sys
 import os
 from re import compile as re, M
 
 import json
 import logging
+
 from tornado import ioloop, log
+import yaml
 
 from .alerts import BaseAlert
 from .utils import parse_interval
 from .handlers import registry
-
-try:
-    import yaml
-except ImportError:
-    yaml = None
 
 
 LOGGER = log.gen_log
@@ -27,7 +25,7 @@ class Reactor(object):
     defaults = {
         'auth_password': None,
         'auth_username': None,
-        'config': 'config.json',
+        'config': None,
         'critical_handlers': ['log', 'smtp'],
         'debug': False,
         'format': 'short',
@@ -61,14 +59,25 @@ class Reactor(object):
         self.callback = ioloop.PeriodicCallback(
             self.repeat, parse_interval(self.options['repeat_interval']))
 
+    def started(self):
+        """Check whether the reactor is running.
+
+        :rtype: bool
+        """
+        return hasattr(self, 'callback') and self.callback.is_running()
+
     def reinit(self, *args, **options):  # pylint: disable=unused-argument
         LOGGER.info('Read configuration')
 
         self.options.update(options)
 
-        self.include_config(self.options.get('config'))
+        config_valid = self.include_config(self.options.get('config'))
         for config in self.options.pop('include', []):
-            self.include_config(config)
+            config_valid = config_valid and self.include_config(config)
+
+        # If we haven't started the ioloop yet and config is invalid then fail fast.
+        if not self.started() and not config_valid:
+            sys.exit(1)
 
         if not self.options['public_graphite_url']:
             self.options['public_graphite_url'] = self.options['graphite_url']
@@ -95,16 +104,22 @@ class Reactor(object):
     def include_config(self, config):
         LOGGER.info('Load configuration: %s' % config)
         if config:
-            loader = yaml.load if yaml and config.endswith('.yml') else json.loads
+            loader_name, loader = _get_loader(config)
+            LOGGER.debug('Using loader: %s' % loader_name)
+            if not loader:
+                return False
             try:
                 with open(config) as fconfig:
-                    source = COMMENT_RE.sub("", fconfig.read())
+                    source = fconfig.read()
+                    if loader_name == 'json':
+                        source = COMMENT_RE.sub("", source)
                     config = loader(source)
                     self.options.get('alerts').extend(config.pop("alerts", []))
                     self.options.update(config)
-
             except (IOError, ValueError):
                 LOGGER.error('Invalid config file: %s' % config)
+                return False
+        return True
 
     def reinit_handlers(self, level='warning'):
         for name in self.options['%s_handlers' % level]:
@@ -172,3 +187,20 @@ def _get_numeric_log_level(level):
         except KeyError:
             raise ValueError("Unknown log level: %s" % level)
     return level
+
+
+def _get_loader(config):
+    """Determine which config file type and loader to use based on a filename.
+
+    :param config str: filename to config file
+    :return: a tuple of the loader type and callable to load
+    :rtype: (str, Callable)
+    """
+    if config.endswith('.yml') or config.endswith('.yaml'):
+        if not yaml:
+            LOGGER.error("pyyaml must be installed to use the YAML loader")
+            # TODO: stop reactor if running
+            return None, None
+        return 'yaml', yaml.load
+    else:
+        return 'json', json.loads